backend/pyproject.toml
[project]
name = "greenlink-backend"
version = "0.1.0"
description = "GreenLink backend foundation"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "passlib>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "pydantic[email]",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.12",
  "redis>=5.1.0",
  "sqlalchemy>=2.0.36",
  "typer>=0.12.5",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.0",
]

[tool.pytest.ini_options]
addopts = "-q -p no:cacheprovider"
testpaths = ["tests"]
cache_dir = ".pytest_cache_local"

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
  ".venv",
  ".uv-cache",
  ".pytest_cache",
  ".pytest_cache_local",
  "pytest-cache-files-*"
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501"]
"app/api/routes/*.py" = ["B008"]
"app/auth/dependencies.py" = ["B008"]
"tests/*.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
